feat: Pila del carrito a Java, dinero a BigDecimal, XSS, responsive y 403
- RF-DS-03: mueve el deshacer del carrito al backend (CarritoUndoService,
  /pedidos/carrito/{push,undo}); elimina js/utils/pila.js
- B-18: precio_actual y precio_unitario a numeric(10,2) + BigDecimal en
  entidades, DTOs y sumas de ventas (db/003)
- B-F-11: escaparHtml en las tablas de usuarios/insumos/productos/dashboard
- B-F-13: api.js maneja 403 con mensaje claro
- responsive: media query <=768px (sidebar como barra superior)
- test de contrato PUT/DELETE /admin/usuarios/{id} (200/403/409)
- 63 tests backend verdes

+++ backend/src/main/java/com/coffee/backend/dto/request/ProductoRequestDTO.java
@@ -5,13 +5,14 @@ import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.PositiveOrZero;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 // Alta/edición de producto (ADMIN). La categoría se referencia por id.
 // `receta` puede venir vacía/null → el producto no consume stock al venderse.
 public record ProductoRequestDTO(
         @NotBlank String nombreProducto,
-        @NotNull @PositiveOrZero Double precioActual,
+        @NotNull @PositiveOrZero BigDecimal precioActual,
         @NotNull Long categoriaId,
         @Valid List<RecetaItemRequestDTO> receta
 ) {
