+++ backend/src/main/java/com/coffee/backend/dto/response/DashboardResponseDTO.java
@@ -1,5 +1,6 @@
 package com.coffee.backend.dto.response;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 /**
@@ -7,7 +8,7 @@ import java.util.List;
  * ventas y pedidos del día, producto más vendido del mes y los insumos con stock bajo.
  */
 public record DashboardResponseDTO(
-        double totalVentasDia,
+        BigDecimal totalVentasDia,
         long totalPedidosDia,
         String productoEstrella,
         List<StockBajoDTO> stockBajo) {
