+++ backend/src/main/java/com/coffee/backend/dto/response/DetalleBoletaResponseDTO.java
@@ -1,8 +1,10 @@
 package com.coffee.backend.dto.response;
 
+import java.math.BigDecimal;
+
 public record DetalleBoletaResponseDTO(
         String nombreProducto,
         Long cantidadPedida,
-        Double precioUnitario
+        BigDecimal precioUnitario
 ) {
 }
